+++ modules/elementary_functions/builtin/include/isscalarBuiltin.hpp
@@ -13,14 +13,11 @@
 #include "Evaluator.hpp"
 #include "nlsElementary_functions_builtin_exports.h"
 //=============================================================================
-namespace Nelson {
-namespace ElementaryFunctionsGateway {
-    //=============================================================================
-    NLSELEMENTARY_FUNCTIONS_BUILTIN_IMPEXP
-    ArrayOfVector
-    isscalarBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
-    //=============================================================================
-}
+namespace Nelson::ElementaryFunctionsGateway {
+//=============================================================================
+NLSELEMENTARY_FUNCTIONS_BUILTIN_IMPEXP
+ArrayOfVector
+isscalarBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================